+++ src/App.js
@@ -3,7 +3,8 @@ import { createTheme, ThemeProvider } from '@mui/material';
 import { Routes, Route } from 'react-router-dom';
 import Home from './components/views/home';
 import List from './components/views/list';
-import Login from './components/views/login';
+import LogIn from './components/views/login';
+import SignUp from './components/views/signup'
 import AddSource from './components/views/addSource';
 import EditSource from './components/views/editSource';
 import { ProtectedRoute } from './components/protectedRoute';
@@ -12,7 +13,6 @@ import Amplify from '@aws-amplify/core';
 import awsconfig from './aws-exports';
 import './App.css';
 import '@aws-amplify/ui-react/styles.css';
-import SignInSide from './components/views/login2';
 
 Amplify.configure(awsconfig);
 
@@ -80,9 +80,9 @@ function App() {
               </ProtectedRoute>
             }
           />
-          <Route path='/' element={<Login/>} />
-          <Route path='login' element={<Login/>} />
-          <Route path='login2' element={<SignInSide/>} />
+          <Route path='/' element={<LogIn/>} />
+          <Route path='login' element={<LogIn/>} />
+          <Route path='signup' element={<SignUp/>} />
         </Routes>
       </ThemeProvider>
     </div>
